Split classes into their own files
