+++ lib/sendgrid-ruby.rb
@@ -1,4 +1,24 @@
 require_relative 'sendgrid/client'
 require_relative 'sendgrid/version'
+require_relative 'sendgrid/helpers/mail/asm'
+require_relative 'sendgrid/helpers/mail/attachment'
+require_relative 'sendgrid/helpers/mail/bcc_settings'
+require_relative 'sendgrid/helpers/mail/bypass_list_management'
+require_relative 'sendgrid/helpers/mail/category'
+require_relative 'sendgrid/helpers/mail/click_tracking'
+require_relative 'sendgrid/helpers/mail/content'
+require_relative 'sendgrid/helpers/mail/custom_arg'
+require_relative 'sendgrid/helpers/mail/email'
+require_relative 'sendgrid/helpers/mail/footer'
+require_relative 'sendgrid/helpers/mail/ganalytics'
+require_relative 'sendgrid/helpers/mail/header'
 require_relative 'sendgrid/helpers/mail/mail'
+require_relative 'sendgrid/helpers/mail/mail_settings'
+require_relative 'sendgrid/helpers/mail/open_tracking'
+require_relative 'sendgrid/helpers/mail/personalization'
+require_relative 'sendgrid/helpers/mail/section'
+require_relative 'sendgrid/helpers/mail/spam_check'
+require_relative 'sendgrid/helpers/mail/subscription_tracking'
+require_relative 'sendgrid/helpers/mail/substitution'
+require_relative 'sendgrid/helpers/mail/tracking_settings'
 require_relative 'sendgrid/helpers/settings/settings'
